+++ Program.cs
@@ -8,25 +8,22 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using MatterHackers.Agg.UI;
 using MatterHackers.MatterControl.DataStorage;
+using MatterHackers.MatterControl.SettingsManagement;
+using Mindscape.Raygun4Net;
+using MatterHackers.Agg.Platform;
 
 namespace MatterHackers.MatterControl
 {
 	static class Program
 	{
-		
-		// ** Standard Winforms Main ** //
-		//[STAThread]
-		//static void Main()
-		//{
-		//	Application.EnableVisualStyles();
-		//	Application.SetCompatibleTextRenderingDefault(false);
-		//	Application.Run(new Form1());
-		//}
+        private const int RaygunMaxNotifications = 15;
 
-		/// <summary>
-		/// The main entry point for the application.
-		/// </summary>
-		[STAThread]
+        private static int raygunNotificationCount = 0;
+
+        /// <summary>
+        /// The main entry point for the application.
+        /// </summary>
+        [STAThread]
 		public static void Main()
 		{
 			CultureInfo.DefaultThreadCurrentCulture = CultureInfo.InvariantCulture;
@@ -69,5 +66,28 @@ namespace MatterHackers.MatterControl
 			}
 #endif
 		}
-	}
+
+        private static RaygunClient _raygunClient = GetCorrectClient();
+
+        private static RaygunClient GetCorrectClient()
+        {
+            if (AggContext.OperatingSystem == OSType.Mac)
+            {
+                return new RaygunClient("qmMBpKy3OSTJj83+tkO7BQ=="); // this is the Mac key
+            }
+            else
+            {
+                return new RaygunClient("hQIlyUUZRGPyXVXbI6l1dA=="); // this is the PC key
+            }
+        }
+
+        // ** Standard Winforms Main ** //
+        //[STAThread]
+        //static void Main()
+        //{
+        //	Application.EnableVisualStyles();
+        //	Application.SetCompatibleTextRenderingDefault(false);
+        //	Application.Run(new Form1());
+        //}
+    }
 }
